requirements.txt
\
fastapi
uvicorn
pillow
openai
ultralytics
